hackathon/Fam.ai (Novathon)/backend/requirements.txt
fastapi~=0.111.1
uvicorn~=0.30.3
pymongo
pymongo[srv]
motor~=3.4.0
aiofiles~=23.2.1
starlette~=0.37.2
pydantic~=2.7.4
google-generativeai
pyngrok~=7.2.1
imutils~=0.5.4
requests
pillow~=10.4.0
opencv-python~=4.7.0.72
tensorflow~=2.18.0
google~=3.0.0
python-dotenv~=0.16.0
web3
eth-brownie

